feat: route_service 파일elivation계산부, api요청부 분리.
